Moved SelectRepairedTypeWindow into separate file
