+++ Assets/MixedRealityToolkit/Inspectors/PropertyDrawers/SelectRepairedTypeWindow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1816cf73665b5c648a1545d130811675
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 961230b29c294bb780054c5d02eb6180, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
