+++ Assets/MixedRealityToolkit/Inspectors/PropertyDrawers/TypeReferencePropertyDrawer.cs
@@ -356,54 +356,5 @@ namespace Microsoft.MixedReality.Toolkit.Editor
         {
             DrawTypeSelectionControl(position, property.FindPropertyRelative("reference"), label, attribute as SystemTypeAttribute);
         }
-
-        #region popup window definition
-
-        public class SelectRepairedTypeWindow : EditorWindow
-        {
-            private static Type[] repairedTypeOptions;
-            private static SerializedProperty property;
-            private static SelectRepairedTypeWindow window;
-
-            public static bool WindowOpen { get { return window != null; } }
-
-            public static void Display (Type[] repairedTypeOptions, SerializedProperty property)
-            {
-                if (window != null)
-                    window.Close();
-
-                SelectRepairedTypeWindow.repairedTypeOptions = repairedTypeOptions;
-                SelectRepairedTypeWindow.property = property;
-
-                window = ScriptableObject.CreateInstance(typeof(SelectRepairedTypeWindow)) as SelectRepairedTypeWindow;
-                window.titleContent = new GUIContent("Select repaired type");
-                window.ShowUtility();
-            }
-
-            private void OnGUI()
-            {
-                for (int i = 0; i < repairedTypeOptions.Length; i++)
-                {
-                    if (GUILayout.Button(repairedTypeOptions[i].FullName, EditorStyles.miniButton))
-                    {
-                        property.stringValue = SystemType.GetReference(repairedTypeOptions[i]);
-                        property.serializedObject.ApplyModifiedProperties();
-                        Close();
-                    }
-                }
-            }
-
-            private void OnDisable()
-            {
-                window = null;
-            }
-
-            private void OnInspectorUpdate()
-            {
-                Repaint();
-            }
-        }
-
-        #endregion
     }
 }
